Create TranscriptionTableRow Component
